StreakUp/backend/requirements.txt
Flask==3.1.3
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.48
python-dotenv==1.2.2
psycopg[binary]==3.3.3
gunicorn==25.3.0
openai==2.30.0
